requirements.txt
pip==23.0.1
python-dotenv==0.21.1
requests==2.28.2
pycountry==22.3.5
geopy==2.3.0
folium==0.14.0
Flask==2.2.3